add brave command

+++ start.sh
@@ -1,10 +1,18 @@
 #!/usr/bin/env bash
 # Turn on screen
 # kill script if already running
-kill -9 $(<arrivals_pid.txt)
-rm arrivals_pid.txt
+if [ -f arrivals_pid.txt ]; then
+    kill -9 $(<arrivals_pid.txt)
+    rm arrivals_pid.txt
+fi
 # Close browser
-# Open browser
+if [ -f brave_pid.txt ]; then
+    kill $(<brave_pid.txt)
+    rm brave_pid.txt
+fi
+# Start browser storing PID
+nohup brave-browser --suppress-message-center-popups --start-fullscreen file:///home/ubuntu/Documents/HillelDepartureBoard/DepartureBoard.html >/dev/null 2>&1 &
+echo $! > brave_pid.txt
 # Start arrivals script storing PID
 nohup ./arrivals.py --marc_code 12018-12015 --metro_code E09 --refresh 20 >/dev/null 2>&1 &
 echo $! > arrivals_pid.txt
